# api.yaml
openapi: 3.0.1
info:
  title: Clip Rate
  contact:
    email: hello+bt+cliprate@bweston.me
  license:
    name: MIT
  version: 1.0.0
...
servers:
- url: http://localhost:8000/v1
...
paths:
...
  /login:
    post:
      tags:
      - auth
      summary: Login with Twitch code.
      operationId: loginWithTwitchCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/loginWithTwitchCodeRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/loginWithTwitchCodeResponse'
        401:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
    loginWithTwitchCodeRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
    loginWithTwitchCodeResponse:
      type: object
      required:
        - token
        - self
      properties:
        token:
          type: string
        self:
          $ref: '#/components/schemas/Subscriber'
...
    Subscriber:
      type: object
      required:
        - subscriber_id
        - email
        - display_name
        - twitch_id
        - joined
        - is_subscribed
        - is_admin
      properties:
        subscriber_id:
          type: string
          description: Our subscriber identifier.

        email:
          type: string
          description: Email address of the subscriber, this is where notifications would be sent too.

        display_name:
          type: string
          description: The display name of the subscriber.

        twitch_id:
          type: string
          description: Twitch's identifier for the subscriber.

        joined:
          type: string
          description: A timestamp for when the subscriber logged into this application first.

        is_subscribed:
          type: boolean
          description: Is the subscriber currently subscriber.

        is_admin:
          type: boolean
          description: Is the subscriber a mod or owner of the broadcaster.
    ApiResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header