# app/src/api/models/loginWithTwitchCodeResponse.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Subscriber } from './Subscriber';

export type loginWithTwitchCodeResponse = {
    token: string;
    self: Subscriber;
}